backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
openai==1.12.0
anthropic==0.18.1
